fix: impove option implementation
